Adds checking for admin user to the server
