+++ app/entry.server.tsx
@@ -1,40 +1,15 @@
 import { PassThrough } from "node:stream";
 
-import type { AppLoadContext, EntryContext } from "react-router";
 import { createReadableStreamFromReadable } from "@react-router/node";
-import { ServerRouter } from "react-router";
 import { isbot } from "isbot";
 import type { RenderToPipeableStreamOptions } from "react-dom/server";
 import { renderToPipeableStream } from "react-dom/server";
+import type { AppLoadContext, EntryContext } from "react-router";
+import { ServerRouter } from "react-router";
 
-import './modules/storage/storage';
 import './modules/documents/documents';
-import getDocumentsAdapter from "./modules/documents/helpers/getDocumentsAdapter";
-import type { User } from "./modules/users/users.types";
 import createQueue from "./modules/queues/helpers/createQueue";
-
-const checkSuperAdminExists = async () => {
-  const documents = getDocumentsAdapter();
-
-  const user = await documents.getDocument({
-    collection: 'users',
-    match: { role: 'SUPER_ADMIN', githubId: parseInt(process.env.SUPER_ADMIN_GITHUB_ID as string) }
-  }) as { data: User | undefined };
-
-  if (!user.data) {
-    await documents.createDocument({
-      collection: 'users',
-      update: {
-        role: 'SUPER_ADMIN',
-        username: 'local',
-        githubId: process.env.SUPER_ADMIN_GITHUB_ID,
-        hasGithubSSO: process.env.SUPER_ADMIN_GITHUB_ID ? true : false,
-        isRegistered: true,
-        registeredAt: new Date()
-      }
-    })
-  }
-}
+import './modules/storage/storage';
 
 const setupQueues = async () => {
   createQueue('tasks');
@@ -42,7 +17,6 @@ const setupQueues = async () => {
 }
 
 setTimeout(() => {
-  checkSuperAdminExists();
   setupQueues();
 }, 0);
